+++ packages/streamid/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.2.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/streamid@3.1.0...@ceramicnetwork/streamid@3.2.0-rc.0) (2023-12-07)
+
+**Note:** Version bump only for package @ceramicnetwork/streamid
+
+
+
+
+
+# [3.1.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/streamid@3.1.0-rc.0...@ceramicnetwork/streamid@3.1.0) (2023-12-06)
+
+**Note:** Version bump only for package @ceramicnetwork/streamid
+
+
+
+
+
 # [3.1.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/streamid@3.0.4...@ceramicnetwork/streamid@3.1.0-rc.0) (2023-12-01)
 
 **Note:** Version bump only for package @ceramicnetwork/streamid